requirements.txt
streamlit~=1.42.0
google-genai~=1.2.0
pillow~=11.1.0
pydantic~=2.9.2